+++ src/components/keyboard.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 use yew::prelude::*;
 
-use crate::state::{TileState, GameMode};
+use crate::manager::{TileState, GameMode};
 use crate::Msg;
 
 use crate::components::{message::Message};
@@ -17,6 +17,7 @@ pub struct Props {
     pub is_unknown: bool,
     pub is_winner: bool,
     pub is_guessing: bool,
+    pub is_hidden: bool,
     pub game_mode: GameMode,
 
     pub message: String,
@@ -41,6 +42,7 @@ pub fn keyboard(props: &Props) -> Html {
                 is_unknown={props.is_unknown}
                 is_winner={props.is_winner}
                 is_guessing={props.is_guessing}
+                is_hidden={props.is_hidden}
                 last_guess={props.last_guess.clone()}
                 word={props.word.clone()}
                 game_mode={props.game_mode}
@@ -56,8 +58,14 @@ pub fn keyboard(props: &Props) -> Html {
                             callback.emit(Msg::KeyPress(*key));
                         });
 
+                        let tile_state = if !props.is_hidden {
+                            props.keyboard.get(key).unwrap().to_string()
+                        } else {
+                            TileState::Unknown.to_string()
+                        };
+
                         html! {
-                            <button data-nosnippet="" class={classes!("keyboard-button", props.keyboard.get(key).unwrap().to_string())}
+                            <button data-nosnippet="" class={classes!("keyboard-button", tile_state)}
                                 onmousedown={onkeypress}>
                                 { key }
                             </button>
@@ -78,8 +86,14 @@ pub fn keyboard(props: &Props) -> Html {
                             callback.emit(Msg::KeyPress(*key));
                         });
 
+                        let tile_state = if !props.is_hidden {
+                            props.keyboard.get(key).unwrap().to_string()
+                        } else {
+                            TileState::Unknown.to_string()
+                        };
+
                         html! {
-                            <button data-nosnippet="" class={classes!("keyboard-button", props.keyboard.get(key).unwrap().to_string())}
+                            <button data-nosnippet="" class={classes!("keyboard-button", tile_state)}
                                 onmousedown={onkeypress}>
                                 { key }
                             </button>
@@ -99,8 +113,14 @@ pub fn keyboard(props: &Props) -> Html {
                             callback.emit(Msg::KeyPress(*key));
                         });
 
+                        let tile_state = if !props.is_hidden {
+                            props.keyboard.get(key).unwrap().to_string()
+                        } else {
+                            TileState::Unknown.to_string()
+                        };
+
                         html! {
-                            <button data-nosnippet="" class={classes!("keyboard-button", props.keyboard.get(key).unwrap().to_string())}
+                            <button data-nosnippet="" class={classes!("keyboard-button", tile_state)}
                                 onmousedown={onkeypress}>{ key }</button>
                         }
                     }).collect::<Html>()
@@ -119,7 +139,7 @@ pub fn keyboard(props: &Props) -> Html {
                                 { "ARVAA" }
                             </button>
                         }
-                    } else if let GameMode::DailyWord(_) = props.game_mode {
+                    } else if matches!(props.game_mode, GameMode::DailyWord(_) | GameMode::Shared) {
                         let callback = props.callback.clone();
                         let onmousedown = Callback::from(move |e: MouseEvent| {
                             e.prevent_default();