+++ src/game.rs
@@ -1,421 +1,23 @@
 use rand::seq::SliceRandom;
 use std::collections::HashMap;
-use std::collections::HashSet;
-use std::fmt;
 use std::mem;
 use std::rc::Rc;
 
-use chrono::{Local, NaiveDate};
+use chrono::NaiveDate;
 use gloo_storage::{errors::StorageError, LocalStorage, Storage};
 use serde::{Deserialize, Serialize};
+use web_sys::{window, Window};
 
-use crate::migration;
+use crate::manager::{
+    CharacterCount, CharacterState, GameMode, Theme, TileState, WordList, WordLists,
+};
 
-const FULL_WORDS: &str = include_str!("../full-words.txt");
-const COMMON_WORDS: &str = include_str!("../common-words.txt");
 const DAILY_WORDS: &str = include_str!("../daily-words.txt");
-const PROFANITIES: &str = include_str!("../profanities.txt");
 const SUCCESS_EMOJIS: [&str; 8] = ["🥳", "🤩", "🤗", "🎉", "😊", "😺", "😎", "👏"];
 pub const EMPTY: char = '\u{00a0}'; // &nbsp;
 pub const DEFAULT_WORD_LENGTH: usize = 5;
 pub const DEFAULT_MAX_GUESSES: usize = 6;
 pub const DEFAULT_ALLOW_PROFANITIES: bool = false;
-pub const DAILY_WORD_LEN: usize = 5;
-
-type WordLists = HashMap<(WordList, usize), HashSet<Vec<char>>>;
-
-fn parse_all_words() -> Rc<WordLists> {
-    let mut word_lists: HashMap<(WordList, usize), HashSet<Vec<char>>> = HashMap::with_capacity(3);
-    for word in FULL_WORDS.lines() {
-        let chars = word.chars();
-        let word_length = chars.clone().count();
-        word_lists
-            .entry((WordList::Full, word_length))
-            .or_insert_with(HashSet::new)
-            .insert(chars.collect());
-    }
-
-    for word in COMMON_WORDS.lines() {
-        let chars = word.chars();
-        let word_length = chars.clone().count();
-        word_lists
-            .entry((WordList::Common, word_length))
-            .or_insert_with(HashSet::new)
-            .insert(chars.collect());
-    }
-
-    for word in PROFANITIES.lines() {
-        let chars = word.chars();
-        let word_length = chars.clone().count();
-        word_lists
-            .entry((WordList::Profanities, word_length))
-            .or_insert_with(HashSet::new)
-            .insert(chars.collect());
-    }
-
-    Rc::new(word_lists)
-}
-
-#[derive(PartialEq, Eq, Hash, Clone, Copy, Serialize, Deserialize)]
-pub enum WordList {
-    Full,
-    Common,
-    Profanities,
-    Daily,
-}
-
-impl Default for WordList {
-    fn default() -> Self {
-        WordList::Common
-    }
-}
-
-#[derive(PartialEq, Eq, Hash, Clone, Copy, Serialize, Deserialize)]
-pub enum GameMode {
-    Classic,
-    Relay,
-    DailyWord(NaiveDate),
-}
-
-impl Default for GameMode {
-    fn default() -> Self {
-        GameMode::Classic
-    }
-}
-
-#[derive(PartialEq, Clone, Copy, Serialize, Deserialize)]
-pub enum Theme {
-    Dark,
-    Colorblind,
-}
-
-impl Default for Theme {
-    fn default() -> Self {
-        Theme::Dark
-    }
-}
-
-#[derive(Clone, PartialEq, Serialize, Deserialize)]
-pub enum CharacterState {
-    Correct,
-    Absent,
-    Unknown,
-}
-
-#[derive(Clone, PartialEq, Serialize, Deserialize)]
-pub enum TileState {
-    Correct,
-    Absent,
-    Present,
-    Unknown,
-}
-
-impl fmt::Display for TileState {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            TileState::Correct => write!(f, "correct"),
-            TileState::Absent => write!(f, "absent"),
-            TileState::Present => write!(f, "present"),
-            TileState::Unknown => write!(f, "unknown"),
-        }
-    }
-}
-
-#[derive(Clone, PartialEq, Serialize, Deserialize)]
-pub enum CharacterCount {
-    AtLeast(usize),
-    Exactly(usize),
-}
-
-#[derive(Clone, PartialEq, Serialize, Deserialize)]
-pub struct State {
-    pub allow_profanities: bool,
-    pub current_game_mode: GameMode,
-    pub current_word_list: WordList,
-    pub current_word_length: usize,
-    pub current_max_guesses: usize,
-
-    pub previous_game: (GameMode, WordList, usize),
-
-    pub theme: Theme,
-
-    pub max_streak: usize,
-    pub total_played: usize,
-    pub total_solved: usize,
-
-    #[serde(skip)]
-    pub game: Game,
-    #[serde(skip)]
-    pub background_games: HashMap<(GameMode, WordList, usize), Game>,
-    #[serde(skip)]
-    pub word_lists: Rc<WordLists>,
-}
-
-impl Default for State {
-    fn default() -> Self {
-        Self {
-            allow_profanities: DEFAULT_ALLOW_PROFANITIES,
-            current_game_mode: GameMode::default(),
-            current_word_list: WordList::default(),
-            current_word_length: DEFAULT_WORD_LENGTH,
-            current_max_guesses: DEFAULT_MAX_GUESSES,
-
-            previous_game: (
-                GameMode::default(),
-                WordList::default(),
-                DEFAULT_WORD_LENGTH,
-            ),
-
-            theme: Theme::default(),
-
-            max_streak: 0,
-            total_played: 0,
-            total_solved: 0,
-
-            game: Game::default(),
-            background_games: HashMap::new(),
-            word_lists: Rc::new(HashMap::new()),
-        }
-    }
-}
-
-impl State {
-    pub fn new() -> Self {
-        let word_lists = parse_all_words();
-
-        // Attempt to rehydrate state from localStorage
-        if let Ok(mut state) = State::rehydrate() {
-            if let GameMode::DailyWord(date) = state.current_game_mode {
-                let today = Local::today().naive_local();
-
-                if date < today {
-                    // Page was refreshed after the day changed - rehydrate the daily word of today
-                    state.current_game_mode = GameMode::DailyWord(today);
-                }
-            }
-
-            let game = Game::new_or_rehydrate(
-                state.current_game_mode,
-                state.current_word_list,
-                state.current_word_length,
-                state.allow_profanities,
-                word_lists.clone(),
-            );
-
-            state.game = game;
-            state.word_lists = word_lists;
-
-            state
-        } else {
-            // Otherwise either create everything from scratch or recover some data from legacy storage state
-            let game = Game::new(
-                GameMode::Classic,
-                WordList::Common,
-                DEFAULT_WORD_LENGTH,
-                DEFAULT_ALLOW_PROFANITIES,
-                word_lists.clone(),
-            );
-
-            let mut state = Self {
-                game,
-                word_lists,
-                ..Self::default()
-            };
-
-            // Try to migrate old settings and stats from localStorage to current format
-            // TODO: Doesn't do anything if the old state isn't present, but get rid of this at some point
-            let _res = migration::migrate_settings_and_stats(&mut state);
-            state.switch_active_game();
-
-            // Try to migrate old game streak from localStorage to current format, if the game mode is not daily
-            // TODO: Doesn't do anything if the old state isn't present, but get rid of this at some point
-            let _res = migration::migrate_game(&mut state.game);
-
-            let _res = state.persist();
-            let _res = state.game.persist();
-
-            state
-        }
-    }
-
-    pub fn submit_guess(&mut self) -> bool {
-        self.game.submit_guess();
-        if !self.game.is_guessing {
-            self.update_game_statistics(self.game.is_winner, self.game.streak);
-        }
-
-        true
-    }
-
-    pub fn change_word_length(&mut self, new_length: usize) {
-        if self.current_word_length == new_length {
-            return;
-        }
-
-        self.current_word_length = new_length;
-        self.switch_active_game();
-
-        let _res = self.persist();
-        let _res = self.game.persist();
-    }
-
-    pub fn change_game_mode(&mut self, new_mode: GameMode) {
-        if self.current_game_mode == new_mode {
-            return;
-        }
-
-        if matches!(self.current_game_mode, GameMode::DailyWord(_)) {
-            self.current_word_list = self.previous_game.1;
-            self.current_word_length = self.previous_game.2;
-        }
-
-        if matches!(new_mode, GameMode::DailyWord(_)) {
-            self.current_word_list = WordList::Daily;
-            self.current_word_length = DAILY_WORD_LEN;
-        } else if self.current_word_list == WordList::Daily {
-            // Prevent getting stuck in non-daily word gamemode with
-            // daily list somehow, for instance by having a daily game as
-            // the previous game in state
-            self.current_word_list = WordList::default();
-        }
-
-        self.current_game_mode = new_mode;
-        self.switch_active_game();
-        let _res = self.persist();
-        let _res = self.game.persist();
-    }
-
-    pub fn change_word_list(&mut self, new_list: WordList) {
-        if self.current_word_list == new_list {
-            return;
-        }
-
-        self.current_word_list = new_list;
-        self.switch_active_game();
-
-        let _res = self.persist();
-        let _res = self.game.persist();
-    }
-
-    pub fn change_previous_game_mode(&mut self) {
-        let (game_mode, word_list, word_length) = self.previous_game;
-
-        if matches!(game_mode, GameMode::DailyWord(_)) && matches!(self.current_game_mode, GameMode::DailyWord(_)) {
-            // Force the user to reset to the base game
-            self.current_game_mode = GameMode::default();
-            self.current_word_list = WordList::default();
-            self.current_word_length = DEFAULT_WORD_LENGTH;
-        } else {
-            self.current_game_mode = game_mode;
-            self.current_word_list = word_list;
-            self.current_word_length = word_length;
-        }
-
-        self.switch_active_game();
-
-        let _res = self.persist();
-        let _res = self.game.persist();
-    }
-
-    pub fn change_allow_profanities(&mut self, is_allowed: bool) {
-        self.allow_profanities = is_allowed;
-        self.game.allow_profanities = self.allow_profanities;
-        self.background_games.values_mut().for_each(|game| {
-            game.allow_profanities = self.allow_profanities;
-        });
-        let _result = self.persist();
-    }
-
-    pub fn change_theme(&mut self, theme: Theme) -> bool {
-        self.theme = theme;
-        let _result = self.persist();
-        true
-    }
-
-    fn switch_active_game(&mut self) -> bool {
-        let next_game = (
-            self.current_game_mode,
-            self.current_word_list,
-            self.current_word_length,
-        );
-
-        let previous_game = (
-            self.game.game_mode,
-            self.game.word_list,
-            self.game.word_length,
-        );
-
-        if next_game.0 == previous_game.0
-            && next_game.1 == previous_game.1
-            && next_game.2 == previous_game.2
-        {
-            return false;
-        }
-
-        self.previous_game = previous_game;
-
-        // Restore a suspended game or create a new one
-        let mut game = self.background_games.remove(&next_game).unwrap_or_else(|| {
-            Game::new_or_rehydrate(
-                next_game.0,
-                next_game.1,
-                next_game.2,
-                self.allow_profanities,
-                self.word_lists.clone(),
-            )
-        });
-
-        // For playing the animation populate previous_guesses
-        if previous_game.2 <= next_game.2 {
-            game.previous_guesses = self.game.guesses.clone();
-        } else {
-            game.previous_guesses = self
-                .game
-                .guesses
-                .iter()
-                .cloned()
-                .map(|guess| guess.into_iter().take(game.word_length).collect())
-                .collect();
-        }
-
-        if self.game.current_guess < game.max_guesses - 1 {
-            game.previous_guesses.truncate(self.game.current_guess);
-        }
-        game.is_reset = true;
-
-        self.background_games
-            .insert(previous_game, mem::replace(&mut self.game, game));
-
-        true
-    }
-
-    fn update_game_statistics(&mut self, is_winner: bool, streak: usize) {
-        self.total_played += 1;
-
-        if is_winner {
-            self.total_solved += 1;
-
-            if streak > self.max_streak {
-                self.max_streak = streak;
-            }
-        }
-    }
-
-    #[cfg(web_sys_unstable_apis)]
-    pub fn share_emojis(&self) -> String {
-        self.game.share_emojis(self.theme)
-    }
-
-    fn persist(&self) -> Result<(), StorageError> {
-        LocalStorage::set("settings", self)
-    }
-
-    fn rehydrate() -> Result<Self, StorageError> {
-        let mut state: Self = LocalStorage::get("settings")?;
-        state.word_lists = parse_all_words();
-        Ok(state)
-    }
-}
 
 #[derive(Clone, PartialEq, Serialize, Deserialize)]
 pub struct Game {
@@ -435,6 +37,8 @@ pub struct Game {
     pub is_reset: bool,
     pub message: String,
 
+    #[serde(skip)]
+    pub is_hidden: bool,
     #[serde(skip)]
     pub allow_profanities: bool,
     #[serde(skip)]
@@ -506,6 +110,7 @@ impl Game {
             is_winner: false,
             is_unknown: false,
             is_reset: false,
+            is_hidden: false,
             message: EMPTY.to_string(),
             known_states,
             discovered_counts,
@@ -516,7 +121,63 @@ impl Game {
         }
     }
 
-    fn new_or_rehydrate(
+    pub fn from_shared_link(game_str: &str, word_lists: Rc<WordLists>) -> Option<Self> {
+        let max_guesses = DEFAULT_MAX_GUESSES;
+
+        let mut parts = game_str.split("|");
+        let word = parts.next()?.chars().collect::<Vec<_>>();
+        let word_length = word.len();
+
+        let guesses_str = parts.next()?;
+
+        let mut guesses = guesses_str
+            .chars()
+            .map(|c| (c, TileState::Unknown))
+            .collect::<Vec<_>>()
+            .chunks(word_length)
+            .map(|chunk| chunk.to_vec())
+            .collect::<Vec<_>>();
+
+        let current_guess = guesses.len() - 1;
+
+        guesses.resize(max_guesses, Vec::with_capacity(word_length));
+
+        let known_states = std::iter::repeat(HashMap::new())
+            .take(max_guesses)
+            .collect::<Vec<_>>();
+
+        let discovered_counts = std::iter::repeat(HashMap::new())
+            .take(max_guesses)
+            .collect::<Vec<_>>();
+
+        let mut game = Self {
+            game_mode: GameMode::Shared,
+            word_list: WordList::Full,
+            word_lists,
+            word_length,
+            max_guesses,
+            word,
+            allow_profanities: true,
+            is_guessing: false,
+            is_winner: false,
+            is_unknown: false,
+            is_reset: false,
+            is_hidden: true,
+            message: EMPTY.to_string(),
+            known_states,
+            discovered_counts,
+            guesses,
+            previous_guesses: Vec::new(),
+            current_guess,
+            streak: 0,
+        };
+
+        game.recalculate();
+
+        return Some(game);
+    }
+
+    pub fn new_or_rehydrate(
         game_mode: GameMode,
         word_list: WordList,
         word_length: usize,
@@ -858,15 +519,18 @@ impl Game {
     }
 
     fn is_guess_real_word(&self) -> bool {
-        match self.word_lists.get(&(WordList::Full, self.word_length)) {
-            Some(list) => {
-                let word: &Vec<char> = &self.guesses[self.current_guess]
-                    .iter()
-                    .map(|(c, _)| *c)
-                    .collect();
+        // Always allow correct words, even if they aren't on the list
+        if self.is_correct_word() {
+            return true;
+        }
 
-                list.contains(word)
-            }
+        let word: &Vec<char> = &self.guesses[self.current_guess]
+            .iter()
+            .map(|(c, _)| *c)
+            .collect();
+
+        match self.word_lists.get(&(WordList::Full, self.word_length)) {
+            Some(list) => list.contains(word),
             None => false,
         }
     }
@@ -906,7 +570,7 @@ impl Game {
         }
     }
 
-    fn submit_guess(&mut self) {
+    pub fn submit_guess(&mut self) {
         if !self.is_guess_allowed() {
             self.message = "Liian vähän kirjaimia!".to_owned();
             return;
@@ -925,8 +589,10 @@ impl Game {
         if self.is_game_ended() {
             self.is_guessing = false;
 
-            if let GameMode::DailyWord(_) = self.game_mode {
-                // Do nothing?
+            if matches!(self.game_mode, GameMode::DailyWord(_))
+                || matches!(self.game_mode, GameMode::Shared)
+            {
+                // Do nothing, don't update streaks
             } else if self.is_winner {
                 self.streak += 1;
             } else {
@@ -937,11 +603,12 @@ impl Game {
         } else {
             self.current_guess += 1;
         }
+
         let _result = self.persist();
     }
 
     #[cfg(web_sys_unstable_apis)]
-    fn share_emojis(&self, theme: Theme) -> String {
+    pub fn share_emojis(&self, theme: Theme) -> String {
         let mut message = String::new();
 
         if let GameMode::DailyWord(date) = self.game_mode {
@@ -983,7 +650,90 @@ impl Game {
         message
     }
 
+    #[cfg(web_sys_unstable_apis)]
+    pub fn share_link(&self) -> Option<String> {
+        let game_str = format!(
+            "{}|{}",
+            self.word.iter().collect::<String>(),
+            self.guesses
+                .iter()
+                .flat_map(|guess| guess.iter().map(|(c, _)| c))
+                .collect::<String>(),
+        );
+        let window: Window = window().expect("window not available");
+        let share_str = window.btoa(&game_str).ok()?;
+        let base_url = window.location().origin().ok()?;
+
+        // Replace +/= at the base64 with URL safe characters
+        let safe_str = share_str.replace("+", "-").replace("/", ".").replace("=", "_");
+
+        Some(format!("{}/?peli={}", base_url, safe_str))
+    }
+
+    pub fn reveal_hidden_tiles(&mut self) -> bool {
+        self.is_hidden = false;
+        self.message = format!("Sana oli \"{}\"", self.word.iter().collect::<String>());
+        true
+    }
+
+    pub fn reset(&mut self) -> bool {
+        self.guesses = std::iter::repeat(Vec::with_capacity(self.word_length))
+            .take(self.max_guesses)
+            .collect::<Vec<_>>();
+
+        self.current_guess = 0;
+
+        self.is_guessing = true;
+        self.is_winner = false;
+        self.is_unknown = false;
+        self.is_reset = false;
+        self.is_hidden = false;
+        self.message = "Peli nollattu, arvaa sanuli!".to_owned();
+
+        self.known_states = std::iter::repeat(HashMap::new())
+            .take(self.max_guesses)
+            .collect::<Vec<_>>();
+
+        self.discovered_counts = std::iter::repeat(HashMap::new())
+            .take(self.max_guesses)
+            .collect::<Vec<_>>();
+
+        self.previous_guesses = Vec::new();
+
+        true
+    }
+
+    fn recalculate(&mut self) {
+        self.known_states = std::iter::repeat(HashMap::new())
+            .take(self.max_guesses)
+            .collect::<Vec<_>>();
+
+        self.discovered_counts = std::iter::repeat(HashMap::new())
+            .take(self.max_guesses)
+            .collect::<Vec<_>>();
+
+        let current_guess = self.current_guess;
+        // Rerun the game to repuplate known_states and discovered_counts
+        for guess_index in 0..self.current_guess {
+            self.current_guess = guess_index;
+            self.calculate_current_guess();
+        }
+
+        // Restore the current guess
+        self.current_guess = current_guess;
+
+        // If the game is ended also recalculate the current guess
+        if !self.is_guessing {
+            self.calculate_current_guess();
+        }
+    }
+
     pub fn persist(&self) -> Result<(), StorageError> {
+        if self.game_mode == GameMode::Shared {
+            // Never persist shared games
+            return Ok(());
+        }
+
         let game_key = &format!(
             "game|{}|{}|{}",
             serde_json::to_string(&self.game_mode).unwrap(),
@@ -1012,28 +762,7 @@ impl Game {
         game.allow_profanities = allow_profanities;
         game.word_lists = word_lists;
 
-        game.known_states = std::iter::repeat(HashMap::new())
-            .take(game.max_guesses)
-            .collect::<Vec<_>>();
-
-        game.discovered_counts = std::iter::repeat(HashMap::new())
-            .take(game.max_guesses)
-            .collect::<Vec<_>>();
-
-        let current_guess = game.current_guess;
-        // Rerrun the game to repuplate known_states and discovered_counts
-        for guess_index in 0..game.current_guess {
-            game.current_guess = guess_index;
-            game.calculate_current_guess();
-        }
-
-        // Restore the current guess
-        game.current_guess = current_guess;
-
-        // If the game is ended also recalculate the current guess
-        if !game.is_guessing {
-            game.calculate_current_guess();
-        }
+        game.recalculate();
 
         Ok(game)
     }